Modified CPGMergeADCs to make it run approximately 1000% faster. The apparent reason for its slowness was that instead of calling TChain::GetTree()->GetEntry() I called TChain::GetEntry() directly. I did this because the grouping of events from different ADC modules took place within Process() where I had to access different input files consecutively. CPGMergeADCs is now written as follows:   * read info_time and module number in Process()   * if synchronized: adjust info_time and add to systimes list   * make a copy of input_chain in SlaveTerminate()   * Parse input branches and link input_branches to the copy   * group events using systimes list and TTree::GetEntry() of the copy Also the source file seems not to be as much of an unreadable havoc as before.
The wrapper however still is of inferior design and needs more rewriting: as of
now, new GEnv items are expected to be added in SlaveTerminate(), and items of
the input file are not available in there, so no CPGConfCal. To make this work
in an intuitive way, Settings.cxx probably has to work with two sepeate TEnv's
- one for new outgoing items and one for all the input items. Since I am the
only one developing reduce-mappers, I'll resolve this issue as soon as I need
to. Just don't expect 'reduce' to work with every mapper.